[speech] Made speech class that cache audio files
Made a Speech class that will cache the audio files, and let the
program continue processing while playing the audio. When a new
file is requested to be played the old one is stopped and a the
file is played.

NOTE: I have not tested this for compile or runtime errors